candle_transformers::models::stable_diffusion

Module uni_pc

UniPC Scheduler

UniPC is a training-free framework designed for the fast sampling of diffusion models, which consists of a corrector (UniC) and a predictor (UniP) that share a unified analytical form and support arbitrary orders.

UniPC is by design model-agnostic, supporting pixel-space/latent-space DPMs on unconditional/conditional sampling. It can also be applied to both noise prediction and data prediction models. Compared with prior methods, UniPC converges faster thanks to the increased order of accuracy. Both quantitative and qualitative results show UniPC can improve sampling quality, especially at very low step counts (5~10).

For more information, see the original publication: UniPC: A Unified Predictor-Corrector Framework for Fast Sampling of Diffusion Models, W. Zhao et al, 2023. https://arxiv.org/abs/2302.04867

This work is based largely on UniPC implementation from the diffusers python package: https://raw.githubusercontent.com/huggingface/diffusers/e8aacda762e311505ba05ae340af23b149e37af3/src/diffusers/schedulers/scheduling_unipc_multistep.py

Structs

EdmDpmMultistepScheduler

ExponentialSigmaSchedule

KarrasSigmaSchedule

UniPCSchedulerConfig

Enums

AlgorithmType

CorrectorConfiguration

FinalSigmasType

SigmaSchedule

SolverType

TimestepSchedule