candle_transformers::models

Module whisper

Whisper Model Implementation

Whisper is an automatic speech recognition (ASR) system trained on large amounts of multilingual and multitask supervised data collected from the web. It can be used to convert audio files (in the .wav format) to text. Supported features include language detection as well as multilingual speech recognition.

⚡ Interactive Wasm Example

💻 GH Link

💻 Transformers Python reference implementation

Modules

audio

model

quantized_model

Structs

Config

Constants

CHUNK_LENGTH

COMPRESSION_RATIO_THRESHOLD

DTYPE

EOT_TOKEN

HOP_LENGTH

LOGPROB_THRESHOLD

NO_SPEECH_THRESHOLD

NO_SPEECH_TOKENS

NO_TIMESTAMPS_TOKEN

N_FFT

N_FRAMES

N_SAMPLES

SAMPLE_RATE

SOT_TOKEN

TEMPERATURES

TRANSCRIBE_TOKEN

TRANSLATE_TOKEN